+++ SamplePlugin/BuildInfo.cs
@@ -2,6 +2,6 @@ namespace SamplePlugin;
 
 public static class BuildInfo
 {
-    public const string Version = "0.3.2";
-    public const string Milestone = "confirmed-territory-1116";
+    public const string Version = "0.3.3";
+    public const string Milestone = "sprint-recommendations";
 }